[System.Reflection] CoreFX import for PropertyInfo type.
