+++ mono/metadata/icall-def.h
@@ -656,6 +656,7 @@ HANDLES(ICALL(MMETHI_3, "get_retval_marshal", ves_icall_System_MonoMethodInfo_ge
 
 ICALL_TYPE(MPROP, "System.Reflection.MonoProperty", MPROP_1)
 HANDLES(ICALL(MPROP_1, "get_metadata_token", ves_icall_reflection_get_token))
+HANDLES(ICALL(MPROP_2, "internal_from_handle_type", ves_icall_System_Reflection_PropertyInfo_internal_from_handle_type))
 
 ICALL_TYPE(MPROPI, "System.Reflection.MonoPropertyInfo", MPROPI_1)
 HANDLES(ICALL(MPROPI_1, "GetTypeModifiers", ves_icall_MonoPropertyInfo_GetTypeModifiers))
@@ -665,9 +666,6 @@ HANDLES(ICALL(MPROPI_2, "get_property_info", ves_icall_MonoPropertyInfo_get_prop
 ICALL_TYPE(PARAMI, "System.Reflection.ParameterInfo", PARAMI_1)
 HANDLES(ICALL(PARAMI_1, "GetMetadataToken", ves_icall_reflection_get_token))
 HANDLES(ICALL(PARAMI_2, "GetTypeModifiers", ves_icall_ParameterInfo_GetTypeModifiers))
-
-ICALL_TYPE(PROPI, "System.Reflection.PropertyInfo", PROPI_1)
-HANDLES(ICALL(PROPI_1, "internal_from_handle_type", ves_icall_System_Reflection_PropertyInfo_internal_from_handle_type))
 
 ICALL_TYPE(RTFIELD, "System.Reflection.RtFieldInfo", RTFIELD_1)
 ICALL(RTFIELD_1, "UnsafeGetValue", ves_icall_MonoField_GetValueInternal)
